% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathertweet-package.R
\docType{package}
\name{gathertweet-package}
\alias{gathertweet}
\alias{gathertweet-package}
\title{gathertweet: Gather Tweets from the Command Line with rtweet}
\description{
A command line utility for common tweet gathering
    tasks. Leverages the rtweet package for cron-friendly,
    boilerplate-free tweet collection. Currently in progress and only
    suitable for small search-based tweet collection projects.
}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com}

}
\keyword{internal}
