% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathertweet_actions.R
\name{gathertweet_search}
\alias{gathertweet_search}
\title{gathertweet actions}
\usage{
gathertweet_search(terms, file = "tweets.rds", n = 18000,
  max_id = NULL, since_id = "last", type = "recent",
  include_rts = FALSE, geocode = NULL, `no-parse` = FALSE,
  token = NULL, retryonratelimit = FALSE, quiet = FALSE, ...)
}
\description{
gathertweet actions
}
