% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_io.R
\name{get_user_info}
\alias{get_user_info}
\title{Get user info}
\usage{
get_user_info(tweets = NULL, file = getOption("gathertweet.file",
  "tweets.rds"), dir_profile_images = NULL)
}
\arguments{
\item{file}{The file where tweets are located. The text \code{_users} is
automatically appended to this file name.}
}
\description{
Get user info
}
