% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_gathertweet}
\alias{install_gathertweet}
\title{Install gathertweet exectuable script}
\usage{
install_gathertweet(location = "/usr/local/bin")
}
\arguments{
\item{location}{Where to install the gathertweet executable script}
}
\description{
Installs the \code{gatherwteet} executable script to the location. Should work
with Unix and MacOS out of the box, but I can't make any guarantees about
Windows.
}
