% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logger}
\alias{logger}
\alias{log_info}
\alias{log_debug}
\alias{log_warn}
\alias{log_error}
\alias{log_fatal}
\alias{log_pipe}
\title{Logging functions}
\usage{
logger(..., level = "info", envir = parent.frame())

log_info(..., envir = parent.frame())

log_debug(..., envir = parent.frame())

log_warn(..., envir = parent.frame())

log_error(..., envir = parent.frame())

log_fatal(..., envir = parent.frame())

log_pipe(.data, ..., level = "info")
}
\description{
Logging functions
}
